feat: implement authentication hooks, character management, and UI components; refactor import paths
